requirements.txt
#add the requirements from the tools you specifically need to use
python-dotenv
requests
pydantic>=2.0
pymupdf
# llama-parse  # optional, uncomment if using LlamaParse backend

# Knowledge graph + vector store
neo4j>=5.0
qdrant-client>=1.7
sentence-transformers>=2.2
pyvis>=0.3
networkx>=3.0

# Testing
pytest>=7.0
